+++ R/plotGram.R
@@ -137,7 +137,7 @@ plotCeps <- function(x, tMin, tMax) {
        length(UIDs) == 0) {
         return(NULL)
     }
-    binData <- getBinaryData(x, UIDs)
+    binData <- getBinaryData(x, UIDs, type='whistle')
     settings <- getPamFft(binData)
     settings$sr <- settings$sr * 2
     settings$wl <- settings$wl * 2
@@ -152,7 +152,7 @@ plotWhistle <- function(x, tMin, tMax) {
        length(UIDs) == 0) {
         return(NULL)
     }
-    binData <- getBinaryData(x, UIDs)
+    binData <- getBinaryData(x, UIDs, type='whistle')
     settings <- getPamFft(binData)
     for(i in seq_along(binData)) {
         plotOneCeps(binData[[i]], hop=settings$hop, sr=settings$sr,