+++ tests/testthat/test-processPps.R
@@ -11,7 +11,6 @@ test_that('Test process database', {
     exPps <- addFunction(exPps, roccaWhistleCalcs, module='WhistlesMoans', verbose=FALSE)
     exPps <- addFunction(exPps, standardCepstrumCalcs, module = 'Cepstrum', verbose=FALSE)
     
-    
     setArgPps <- removeFunction(exPps, 1)
     setArgPps <- addFunction(setArgPps, standardClickCalcs, sr_hz='auto', filterfrom_khz=0, filterto_khz=NULL,
                              winLen_sec=.0025, module='ClickDetector', verbose=FALSE)